Extracts wayback-search WC to separate files
